requirements.txt
requests~=2.31.0
langchain~=0.1.13
python-dotenv~=1.0.1
langchain-openai~=0.1.0
openai~=1.14.2
PyYAML~=6.0.1
qdrant-client~=1.8.2
hnswlib~=0.8.0
tiktoken~=0.6.0